+++ src/app/item-lookup/page.tsx
@@ -125,7 +125,7 @@ export default function Page() {
       </PageHeader>
 
       <div className="flex w-full flex-col items-center">
-        <div className="w-full max-w-3xl">
+        <div className="w-full max-w-4xl">
           <ItemLookupFilters onUpdateFilters={handleUpdateFilters} />
         </div>