+++ src/features/filters/components/parts/ItemCategoryFilters.tsx
@@ -48,12 +48,16 @@ export default function ItemCategoryFilters({
         </div>
 
         <div className="relative flex w-full flex-row items-center shadow-sm">
-          <div className="grid grid-cols-2 gap-x-8 text-left sm:grid-cols-5">
+          <div className="grid grid-cols-2 gap-x-8 text-left sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5">
             {defaultItemCategories.map((category) => {
+              const label =
+                category === 'relicfragment'
+                  ? 'Relic Fragment'
+                  : capitalize(category)
               return (
                 <div key={category}>
                   <Checkbox
-                    label={capitalize(category)}
+                    label={label}
                     name={`category-${category}`}
                     checked={selectedItemCategories.includes(category)}
                     onChange={() => onUpdate(category)}