+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+### 2024-02-09
+
+### Changed
+
+- `relicfragment` updated to read `relic fragment` throughout the site. This is to match the in-game naming convention.
+
 ### 2024-02-08
 
 ### Added